- Minor code changes

+++ src/net/nifheim/beelzebu/coins/core/utils/FileManager.java
@@ -382,10 +382,8 @@ public class FileManager {
         File[] fList = logsFolder.listFiles();
         // Auto purge for old logs
         for (File file : fList) {
-            if (file.isFile() && file.getName().contains(".gz")) {
-                if ((System.currentTimeMillis() - file.lastModified()) >= core.getConfig().getInt("General.Purge.Logs.Days") * 86400000L) {
-                    file.delete();
-                }
+            if (file.isFile() && file.getName().contains(".gz") && (System.currentTimeMillis() - file.lastModified()) >= core.getConfig().getInt("General.Purge.Logs.Days") * 86400000L) {
+                file.delete();
             }
         }
     }
